Import live.2011.09.02.tar.gz
- Updated "MultiFramedRTPSource" to more cleanly handle the situation where the sender's SSRC changes.  We don't yet support
  streams that contain multiple SSRCs, but we can handle streams where the SSRC changes occasionally - e.g., because a multicast
  source has been restarted.  (When the SSRC changes, we don't check the RTP sequence number of the first packet.)
  (Thanks to Dmitriy Vasil'ev for bringing this issue to our attention.)

+++ liveMedia/include/MPEG1or2FileServerDemux.hh
@@ -61,7 +61,7 @@ private:
   Boolean fReuseFirstSource;
   MPEG1or2Demux* fSession0Demux;
   MPEG1or2Demux* fLastCreatedDemux;
-  u_int8_t fLastClientSessionId;
+  unsigned fLastClientSessionId;
 };
 
 #endif
